chore: move all old Meganav folder to the LegacyMeganav so its cleaner
